style: header 스타일 추가
