--- src/components/Navigation/NavigationBar.css
@@ -1,32 +1,0 @@
-.navigation-bar {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 70px;
-  background-color: black;
-  color: #fff;
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  font-weight: bold;
-  padding-right: 20px;
-}
-
-.navigation-bar a {
-  color: white;
-  text-decoration: none;
-  margin-left: 20px;
-}
-
-.page-content {
-  margin-top: 90px; /* 네비게이션 바 높이 + 페이지와 겹치지 않도록 추가 마진 조정 */
-}
-
-.pages {
-  margin-right: 25px;
-}
-
-.toggle {
-  margin-right: 20px;
-}
