--- src/components/Navigation/ToggleBar.css
@@ -1,101 +1,0 @@
-body,
-html {
-  margin: 0;
-  padding: 0;
-  font-family: Arial, sans-serif;
-}
-
-.app {
-  text-align: center;
-  padding: 20px;
-}
-
-.header {
-  background-color: #333;
-  padding: 20px;
-  color: #fff;
-}
-
-.navbar-container {
-  display: flex;
-  justify-content: right;
-}
-
-.navbar {
-  position: relative;
-}
-
-.nav-toggle {
-  cursor: pointer;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 10px;
-}
-
-.bar {
-  width: 30px;
-  height: 4px;
-  background-color: #fff;
-  margin: 4px 0;
-  transition: transform 0.3s ease;
-}
-
-.bar-open {
-  transform: rotate(-45deg) translate(-5px, 6px);
-}
-
-.nav-items {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  opacity: 0;
-  visibility: hidden;
-  position: absolute;
-  top: 100%;
-  left: 50%;
-  transform: translateX(-50%);
-  background-color: #333;
-  border-radius: 4px;
-  transition: all 0.3s ease;
-  flex-direction: column;
-  text-align: center;
-  padding: 0;
-}
-
-.nav-items-show {
-  opacity: 1;
-  visibility: visible;
-  top: 70px;
-  box-shadow: 0px 3px 6px rgba(0, 0, 0, 0.16);
-  left: 50%;
-  transform: translateX(-50%);
-}
-
-.nav-items li {
-  padding: 8px;
-  display: inline;
-  margin: 10px;
-}
-
-.nav-items a {
-  text-decoration: none;
-  color: #fff;
-}
-
-.main {
-  margin-top: 20px;
-}
-
-.footer {
-  background-color: #333;
-  color: #fff;
-  padding: 10px;
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-}
