--- src/components/Page/About.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-const About = () => {
-  return (
-    <div>
-      <h2>소개 페이지입니다. </h2>
-      <p>이곳은 동아리의 소개입니다.</p>
-    </div>
-  );
-};
-
-export default About;
