--- src/components/Page/Board.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Board = () => {
-  return <div className="page-content"> 동아리 게시판 페이지</div>;
-};
-
-export default Board;
