--- src/components/Page/ClubApply.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ClubApply = () => {
-  return <div className="page-content"> 동아리 지원 페이지</div>;
-};
-
-export default ClubApply;
