--- src/components/Page/Contact.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-const Contact = () => {
-  return (
-    <div>
-      <h2>문의 페이지입니다.</h2>
-      <p>이곳은 예쁘고 심플한 랜딩 페이지의 문의 페이지입니다.</p>
-    </div>
-  );
-};
-
-export default Contact;
