--- src/components/Page/Home.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-const Home = () => {
-  return (
-    <div>
-      <h2>안녕하세요! AIDeveloper 홈 페이지입니다.</h2>
-      <p>이곳은 설명을 적을 홈입니다.</p>
-    </div>
-  );
-};
-
-export default Home;
