--- src/components/Page/Introduction.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Introduction = () => {
-  return <div className="page-content"> 동아리 소개 페이지</div>;
-};
-
-export default Introduction;
