--- src/components/Page/Login.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Login = () => {
-  return <div className="page-content"> 동아리 로그인 페이지</div>;
-};
-
-export default Login;
