--- src/components/Page/Main/MainPage.css
@@ -1,48 +1,0 @@
-.typing {
-  height: 55px;
-  display: flex;
-  justify-content: center;
-  align-items: flex-end;
-  font-size: 1.8rem;
-}
-
-.text {
-  font-size: 2.2rem;
-  margin-left: 0.6rem;
-}
-
-.text::after {
-  content: "";
-  margin-left: 0.4rem;
-  border-right: 2px solid #777;
-  animation: cursor 0.9s infinite steps(2);
-}
-
-@keyframes cursor {
-  from {
-    border-right: 2px solid #222;
-  }
-  to {
-    border-right: 2px solid #777;
-  }
-}
-
-/* ---------------------------------------- */
-@import url("https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css");
-
-* {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  box-sizing: border-box;
-  font-family: Pretendard;
-}
-
-body {
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: #222;
-  color: white;
-}
