--- src/components/Page/Main/MainPage.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const MainPage = () => {
-  return (
-    <div className="page-content">
-      Language! <h1 class="text"></h1>
-    </div>
-  );
-};
-
-export default MainPage;
