+++ src/components/base/header/Header.jsx
@@ -1,0 +1,21 @@
+import React from "react";
+import { Link } from "react-router-dom";
+import "./styles/Header.css"; // Import the CSS file
+
+const Header = () => {
+  return (
+    <div className="header">
+      <h1>AID</h1>
+      <h2>elopver</h2>
+
+      {/* <div>
+        <Link to="/intro">동아리 소개</Link>
+        <Link to="/login">로그인</Link>
+        <Link to="/board">게시판</Link>
+        <Link to="/club-apply">동아리 지원</Link>
+      </div> */}
+    </div>
+  );
+};
+
+export default Header;
