+++ src/components/base/sidebar/Sidebar.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+import { Link } from "react-router-dom";
+import "./styles/menu.css";
+
+const Menu = () => {
+  return <div className="Menu">HI</div>;
+};
