+++ src/components/pages/home/home.js
@@ -1,0 +1,3 @@
+import React from "react";
+import { ReactDOM } from "react";
+import "./style/home.css";
