+++ src/index.js
@@ -1,8 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
+import App from "./components/app/App";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
@@ -10,4 +9,3 @@ root.render(
     <App />
   </React.StrictMode>
 );
-reportWebVitals();
